add folder structured
