+++ src/components/Card.js
@@ -1,4 +1,4 @@
-import "./Card.css";
+import "../styles/Card.css";
 
 const Card = ({ id, name, email }) => {
   return (
